+++ nearcore/src/lib.rs
@@ -19,13 +19,14 @@ use near_store::{create_store, Store};
 use near_telemetry::TelemetryActor;
 
 pub use crate::config::{init_configs, load_config, load_test_config, NearConfig, NEAR_BASE};
-use crate::migrations::{migrate_12_to_13, migrate_18_to_19, migrate_19_to_20, migrate_22_to_23};
+use crate::migrations::{
+    migrate_12_to_13, migrate_18_to_19, migrate_19_to_20, migrate_22_to_23, migrate_23_to_24,
+};
 pub use crate::runtime::NightshadeRuntime;
 use near_store::migrations::{
-    fill_col_outcomes_by_hash, fill_col_outcomes_with_metadata, fill_col_transaction_refcount,
-    get_store_version, migrate_10_to_11, migrate_11_to_12, migrate_13_to_14, migrate_14_to_15,
-    migrate_17_to_18, migrate_21_to_22, migrate_6_to_7, migrate_7_to_8, migrate_8_to_9,
-    migrate_9_to_10, set_store_version,
+    fill_col_outcomes_by_hash, fill_col_transaction_refcount, get_store_version, migrate_10_to_11,
+    migrate_11_to_12, migrate_13_to_14, migrate_14_to_15, migrate_17_to_18, migrate_21_to_22,
+    migrate_6_to_7, migrate_7_to_8, migrate_8_to_9, migrate_9_to_10, set_store_version,
 };
 
 #[cfg(feature = "protocol_feature_block_header_v3")]
@@ -218,9 +219,7 @@ pub fn apply_store_migrations(path: &String, near_config: &NearConfig) {
     }
     if db_version <= 23 {
         info!(target: "near", "Migrate DB from version 23 to 24");
-        let store = create_store(&path);
-        fill_col_outcomes_with_metadata(&store);
-        set_store_version(&store, 24);
+        migrate_23_to_24(&path, &near_config);
     }
     #[cfg(feature = "nightly_protocol")]
     {
